docs/ppt-notes-extractor/pyproject.toml
[project]
name = "ppt-notes-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-pptx>=1.0.2",
]
